refactor(lib): extract lib/auth.php — core session + CSRF (ADR-004, sub of #907)
Moves the core session/CSRF/login functions (csrf_token, csrf_require,
require_login, current_user, login_user, logout_user, client_ip, session
lifetime enforcement, …) out of lib.php into lib/auth.php. ADR-003 conversion
applied (global $config -> ipam_config()). Rate-limit, password, and reCAPTCHA
functions stay behind for Tasks 6.2-6.4.
